+++ src/util/Str.c
@@ -87,7 +87,7 @@ char *Str_copy(char *dest, const char *src, int n) {
 char *Str_dup(const char *s) { 
         char *t = NULL;
         if (s) {
-                int n = strlen(s); 
+                size_t n = strlen(s); 
                 t = ALLOC(n + 1);
                 memcpy(t, s, n);
                 t[n] = 0;
@@ -100,7 +100,7 @@ char *Str_ndup(const char *s, int n) {
         char *t = NULL;
         assert(n >= 0);
         if (s) {
-                int l = strlen(s); 
+                int l = (int)strlen(s); 
                 n = l < n ? l : n; // Use the actual length of s if shorter than n
                 t = ALLOC(n + 1);
                 memcpy(t, s, n);
